add footer mobile layout
