+++ .eslintrc.json
@@ -41,7 +41,8 @@
     ],
     "no-unused-vars": ["error", { "argsIgnorePattern": "^_$" }],
     "no-use-before-define": "off",
-    "@typescript-eslint/no-use-before-define": ["error"]
+    "@typescript-eslint/no-use-before-define": ["error"],
+    "@typescript-eslint/ban-ts-comment": "off"
   },
   "settings": {
     "react": {
