+++ webapp/components/Footer/footer.scss
@@ -1,0 +1,47 @@
+@import 'webapp/style/partials';
+
+footer {
+  position: relative;
+  display: grid;
+  align-items: center;
+  grid-template-columns: 1fr auto;
+  grid-column-gap: $spacing-xs;
+  padding: $spacing-xs;
+  border-top: $ui-thin-border;
+  font-size: $font-xxs;
+
+  @include min-width($laptop) {
+    padding: $spacing-s;
+  }
+}
+
+.footer__links {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+}
+
+footer a {
+  text-decoration: none;
+  font-size: $font-xxs - 1;
+  color: $text-body;
+  border-bottom: 1px solid transparent;
+  padding: 1px 0;
+  transition: all 0.25s;
+
+  &:hover {
+    color: $text-body;
+    border-bottom: 1px solid rgba($text-body, 0.5);
+  }
+}
+
+footer .separator {
+  height: $spacing-xxs;
+  width: 1px;
+  background-color: rgba($text-body, 0.7);
+  margin: 0 12px;
+
+  @include min-width($laptop) {
+    height: 12px;
+  }
+}
