--- webapp/components/footer/footer.less
@@ -1,37 +1,0 @@
-@import (reference) 'webapp/main/styles/config';
-
-footer {
-  position: relative;
-  display: flex;
-  align-items: center;
-  padding: 22px 24px;
-  border-top: @ui-thin-border;
-  font-size: 11px;
-}
-
-footer a {
-  text-decoration: none;
-  font-size: 10px;
-  color: @text-body;
-  border-bottom: 1px solid transparent;
-  padding: 1px 0;
-  transition: all 0.25s;
-
-  &:hover {
-    color: @text-body;
-    border-bottom: 1px solid rgba(@text-body, 0.5);
-  }
-}
-
-footer .separator {
-  height: 14px;
-  width: 1px;
-  background-color: rgba(@text-body, 0.7);
-  margin: 0 12px;
-}
-
-footer .copyright {
-  position: absolute;
-  right: 24px;
-  font-size: 11px;
-}
